Restructure log display into same component
